requirements.txt
streamlit
euriai
pymupdf
python-dotenv
apify-client
fastmcp